Docker update for rocBLAS, HCC, and HIP to 2.1 (#1606)
* Updated the requirements with newer HIP, HCC, rocBLAS, and clang-ocl

* Fixed boost issue.

* Fixed tidy issues across the lib.

* Formatting.

* Update hip and rocblas version.

* Fixed remaining tidy issues from formatting.

* Formatting.

* More tidy issues.

* Float conversion.

* NOLINT on stubborn coversion.

* ktime to float

* Remove unneeded NOLINT

* Removed unneeded semi-colon

* Drive updates.

* Switched to rocBLAS 2.3 RC commit hash.

* Use size_t for fusion op_map.


[ROCm/MIOpen commit: ee741f92e5a8e33c9be0556819c3603448363923]

+++ src/solver/conv_ocl_dir2Dfwd_exhaustive_search.cpp
@@ -283,7 +283,7 @@ ConvOclDirectFwdLegacyExhaustiveSearch::SearchImpl(const ConvolutionContext& par
     size_t bot_sz = params.bot_sz / sizeof(Tgpu);
     std::vector<Tgpu> bot_sys_buf(bot_sz);
 
-    for(int i = 0; i < bot_sz; i++)
+    for(size_t i = 0; i < bot_sz; i++)
     {
         bot_sys_buf[i] = static_cast<Tgpu>(rand() * (1.0 / RAND_MAX));
     }
@@ -296,14 +296,14 @@ ConvOclDirectFwdLegacyExhaustiveSearch::SearchImpl(const ConvolutionContext& par
     auto top_ocl_buf = profile_h.Write(top_sys_buf);
 
     std::vector<Tgpu> random_top_sys_buf(top_sz);
-    for(int i = 0; i < top_sz; i++)
+    for(size_t i = 0; i < top_sz; i++)
     {
         random_top_sys_buf[i] = static_cast<Tgpu>(rand() * (1.0 / RAND_MAX));
     }
 
     size_t weights_sz = params.weights_sz / sizeof(Tgpu);
     std::vector<Tgpu> wei_sys_buf(weights_sz);
-    for(int i = 0; i < weights_sz; i++)
+    for(size_t i = 0; i < weights_sz; i++)
     {
         wei_sys_buf[i] = static_cast<Tgpu>((rand() * (1.0 / RAND_MAX) - 0.5) * 0.001);
     }
@@ -317,7 +317,7 @@ ConvOclDirectFwdLegacyExhaustiveSearch::SearchImpl(const ConvolutionContext& par
     {
         size_t bias_sz = params.bias_sz / sizeof(Tgpu);
         bias_sys_buf   = std::vector<Tgpu>(bias_sz);
-        for(int i = 0; i < bias_sz; i++)
+        for(size_t i = 0; i < bias_sz; i++)
         {
             bias_sys_buf[i] = static_cast<Tgpu>(rand() * (1.0 / RAND_MAX));
         }
